backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
mysql-connector-python==8.2.0
PyJWT==2.8.0
